campus-sim/requirements.txt
bacpypes
pymodbus==3.6.9
numpy
netifaces
flask
flask-cors
flask-login
werkzeug
cryptography
websockets
pyyaml
